+++ hydxlib/__init__.py
@@ -1,3 +1,3 @@
 # fmt: off
-__version__ = '1.7.7.dev0'
+__version__ = '1.7.7'
 # fmt: on