+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog of hydxlib
 ===================================================
 
-1.7.7 (unreleased)
+1.7.7 (2026-06-09)
 ------------------
 
 - Set weir cross-section shape to closed rectangle; set cross_section_height from OVS_VOH (mm to m) when present (nens/rana#4429).